Updated Atomist Client to 18.0    - Required updating how we handle config loading. Separated out how the jenkins instances are provisioned.
