+++ src/atomist.config.ts
@@ -38,6 +38,7 @@ import {ProjectCreated} from "./gluon/events/project/ProjectCreated";
 import {ProjectEnvironmentsRequested} from "./gluon/events/project/ProjectEnvironmentsRequested";
 import {TeamsLinkedToProject} from "./gluon/events/project/TeamAssociated";
 import {BotJoinedChannel} from "./gluon/events/team/BotJoinedChannel";
+import {DevOpsEnvironmentProvisioned} from "./gluon/events/team/DevOpsEnvironmentProvisioned";
 import {DevOpsEnvironmentRequested} from "./gluon/events/team/DevOpsEnvironmentRequested";
 import {MembersAddedToTeam} from "./gluon/events/team/MembersAddedToTeam";
 import {MembershipRequestClosed} from "./gluon/events/team/MembershipRequestClosed";
@@ -63,6 +64,8 @@ import {
     SlackIdentity,
 } from "./gluon/ingesters/sharedIngester";
 import {
+    DevOpsEnvironmentDetails,
+    DevOpsEnvironmentProvisionedEvent,
     DevOpsEnvironmentRequestedEvent,
     MembersAddedToTeamEvent,
     MembershipRequestCreatedEvent,
@@ -116,6 +119,7 @@ export const configuration: any = {
         BitbucketProjectRequested,
         BitbucketProjectAdded,
         DevOpsEnvironmentRequested,
+        DevOpsEnvironmentProvisioned,
         ProjectEnvironmentsRequested,
         ApplicationCreated,
         MembershipRequestCreated,
@@ -141,6 +145,8 @@ export const configuration: any = {
         GluonTenant,
         GluonTenantId,
         TeamsLinkedToProjectEvent,
+        DevOpsEnvironmentProvisionedEvent,
+        DevOpsEnvironmentDetails,
     ],
     token,
     http,
