+++ src/config/QMConfig.ts
@@ -1,29 +1,37 @@
-import * as config from "config";
+import fs = require("fs");
 import _ = require("lodash");
 import {HttpAuth} from "./HttpAuth";
 import {SubatomicConfig} from "./SubatomicConfig";
 
 export class QMConfig {
 
-    public static subatomic: SubatomicConfig = config.get("subatomic");
+    public static subatomic: SubatomicConfig;
 
-    public static teamId: string = config.get("teamId");
+    public static teamId: string;
 
-    public static token: string = config.get("token");
+    public static token: string;
 
-    public static http: HttpAuth = config.get("http");
+    public static http: HttpAuth;
 
     public static publicConfig() {
         return new PublicQMConfig();
+    }
+
+    public static initialize() {
+        const config = JSON.parse(fs.readFileSync("config/local.json").toString());
+        QMConfig.subatomic = config.subatomic;
+        QMConfig.teamId = config.teamId;
+        QMConfig.token = config.token;
+        QMConfig.http = config.http;
     }
 
 }
 
 export class PublicQMConfig {
 
-    public subatomic: SubatomicConfig = _.cloneDeep(config.get("subatomic"));
+    public subatomic: SubatomicConfig = _.cloneDeep(QMConfig.subatomic);
 
-    public teamId: string = _.cloneDeep(config.get("teamId"));
+    public teamId: string = _.cloneDeep(QMConfig.teamId);
 
     constructor() {
         this.subatomic.bitbucket.auth.email = "";
@@ -35,3 +43,5 @@ export class PublicQMConfig {
         this.subatomic.openshift.auth.token = "";
     }
 }
+
+QMConfig.initialize();
